requirements.txt
torch
transformers>=4.57.0
peft
accelerate
datasets
pyyaml
pillow
tqdm
